refactor: Restructure audit-worker to standard service architecture
Migrate audit-worker from flat entrypoint + internal/audit-consumer
to the standard service directory structure defined in ADR-0015:

- cmd/: Entry point (main.go moved from root)
- domain/: Core domain models (from internal/audit-consumer/domain)
- adapters/kafka/: Kafka consumer (from internal/audit-consumer/adapters/messaging)
- adapters/persistence/: DB writer (from internal/audit-consumer/adapters/persistence)
- app/: Config and DI container (from internal/audit-consumer/app)
- observability/: Metrics and health checks (from internal/audit-consumer/observability)

Update all import paths across dependent packages:
- cmd/audit-consumer
- services/utilization-metering-consumer
- tests/audit-e2e

Remove internal/audit-consumer/ directory entirely.
All existing tests pass with no functional changes.
