+++ services/audit-worker/adapters/persistence/tenant_audit_writer.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	auditv1 "github.com/meridianhub/meridian/api/proto/meridian/audit/v1"
-	"github.com/meridianhub/meridian/internal/audit-consumer/domain"
+	"github.com/meridianhub/meridian/services/audit-worker/domain"
 	"github.com/meridianhub/meridian/shared/platform/db"
 	"github.com/meridianhub/meridian/shared/platform/tenant"
 	"gorm.io/gorm"
