+++ services/audit-worker/adapters/persistence/tenant_audit_writer_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	auditv1 "github.com/meridianhub/meridian/api/proto/meridian/audit/v1"
-	"github.com/meridianhub/meridian/internal/audit-consumer/adapters/persistence"
+	"github.com/meridianhub/meridian/services/audit-worker/adapters/persistence"
 	"github.com/meridianhub/meridian/shared/platform/tenant"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
