+++ services/utilization-metering-consumer/adapters/messaging/audit_consumer.go
@@ -10,7 +10,7 @@ import (
 
 	"buf.build/go/protovalidate"
 	auditv1 "github.com/meridianhub/meridian/api/proto/meridian/audit/v1"
-	auditdomain "github.com/meridianhub/meridian/internal/audit-consumer/domain"
+	auditdomain "github.com/meridianhub/meridian/services/audit-worker/domain"
 	"github.com/meridianhub/meridian/services/utilization-metering-consumer/domain"
 	"github.com/meridianhub/meridian/shared/platform/kafka"
 	"google.golang.org/protobuf/proto"
