+++ services/utilization-metering-consumer/domain/audit_event_transformer.go
@@ -1,6 +1,6 @@
 // Package domain contains the core domain models for utilization metering.
 //
-// NOTE: This service uses the transformer from internal/audit-consumer/domain
+// NOTE: This service uses the transformer from services/audit-worker/domain
 // instead of maintaining a duplicate implementation. This file is retained
 // only to hold domain-specific interfaces and types that don't belong in
 // the shared internal package.
