+++ tests/audit-e2e/audit_writer_e2e_test.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	auditv1 "github.com/meridianhub/meridian/api/proto/meridian/audit/v1"
-	"github.com/meridianhub/meridian/internal/audit-consumer/adapters/persistence"
+	"github.com/meridianhub/meridian/services/audit-worker/adapters/persistence"
 	"github.com/meridianhub/meridian/shared/platform/tenant"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
